OTel.SDK.Resource.Core.Attributes

Documentation

pattern RESOURCE_SCHEMA_URL :: SchemaURL

The URL of the OpenTelemetry schema for these keys and values.

pattern BROWSER_BRANDS :: Key (AttrVals Text)

Array of brand name and version separated by a space

Notes

This value is intended to be taken from the https://wicg.github.io/ua-client-hints/#interface (navigator.userAgentData.brands).

pattern BROWSER_PLATFORM :: Key Text

The platform on which the browser is running

Notes

This value is intended to be taken from the https://wicg.github.io/ua-client-hints/#interface (navigator.userAgentData.platform). If unavailable, the legacy navigator.platform API SHOULD NOT be used instead and this attribute SHOULD be left unset in order for the values to be consistent. The list of possible values is defined in the https://wicg.github.io/ua-client-hints/#sec-ch-ua-platform. Note that some (but not all) of these values can overlap with values in the ./os.md. However, for consistency, the values in the browser.platform attribute should capture the exact value that the user agent provides.

pattern BROWSER_USER_AGENT :: Key Text

Full user-agent string provided by the browser

Notes

The user-agent value SHOULD be provided only from browsers that do not have a mechanism to retrieve brands and platform individually from the User-Agent Client Hints API. To retrieve the value, the legacy navigator.userAgent API can be used.

pattern CLOUD_PROVIDER :: Key Text

Name of the cloud provider.

pattern CLOUD_ACCOUNT_ID :: Key Text

The cloud account ID the resource is assigned to.

pattern CLOUD_REGION :: Key Text

The geographical region the resource is running.

Notes

Refer to your provider's docs to see the available regions, for example https://www.alibabacloud.com/help/doc-detail/40654.htm, https://aws.amazon.com/about-aws/global-infrastructure/regions_az/, https://azure.microsoft.com/en-us/global-infrastructure/geographies/, https://cloud.google.com/about/locations, or https://intl.cloud.tencent.com/document/product/213/6091.

pattern CLOUD_AVAILABILITY_ZONE :: Key Text

Cloud regions often have multiple, isolated locations known as zones to increase availability. Availability zone represents the zone where the resource is running.

Notes

Availability zones are called &quot;zones&quot; on Alibaba Cloud and Google Cloud.

pattern CLOUD_PLATFORM :: Key Text

The cloud platform in use.

Notes

The prefix of the service SHOULD match the one specified in cloud.provider.

pattern AWS_ECS_CONTAINER_ARN :: Key Text

The Amazon Resource Name (ARN) of an https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ECS_instances.html.

pattern AWS_ECS_CLUSTER_ARN :: Key Text

The ARN of an https://docs.aws.amazon.com/AmazonECS/latest/developerguide/clusters.html.

pattern AWS_ECS_LAUNCHTYPE :: Key Text

The https://docs.aws.amazon.com/AmazonECS/latest/developerguide/launch_types.html for an ECS task.

pattern AWS_ECS_TASK_ARN :: Key Text

The ARN of an https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definitions.html.

pattern AWS_ECS_TASK_FAMILY :: Key Text

The task definition family this task definition is a member of.

pattern AWS_ECS_TASK_REVISION :: Key Text

The revision for this task definition.

pattern AWS_EKS_CLUSTER_ARN :: Key Text

The ARN of an EKS cluster.

pattern AWS_LOG_GROUP_NAMES :: Key (AttrVals Text)

The name(s) of the AWS log group(s) an application is writing to.

Notes

Multiple log groups must be supported for cases like multi-container applications, where a single application has sidecar containers, and each write to their own log group.

pattern AWS_LOG_GROUP_ARNS :: Key (AttrVals Text)

The Amazon Resource Name(s) (ARN) of the AWS log group(s).

Notes

See the https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/iam-access-control-overview-cwl.html#CWL_ARN_Format.

pattern AWS_LOG_STREAM_NAMES :: Key (AttrVals Text)

The name(s) of the AWS log stream(s) an application is writing to.

pattern AWS_LOG_STREAM_ARNS :: Key (AttrVals Text)

The ARN(s) of the AWS log stream(s).

Notes

See the https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/iam-access-control-overview-cwl.html#CWL_ARN_Format. One log group can contain several log streams, so these ARNs necessarily identify both a log group and a log stream.

pattern CONTAINER_NAME :: Key Text

Container name used by container runtime.

pattern CONTAINER_ID :: Key Text

Container ID. Usually a UUID, as for example used to https://docs.docker.com/engine/reference/run/#container-identification. The UUID might be abbreviated.

pattern CONTAINER_RUNTIME :: Key Text

The container runtime managing this container.

pattern CONTAINER_IMAGE_NAME :: Key Text

Name of the image the container was built on.

pattern CONTAINER_IMAGE_TAG :: Key Text

Container image tag.

pattern DEPLOYMENT_ENVIRONMENT :: Key Text

Name of the https://en.wikipedia.org/wiki/Deployment_environment (aka deployment tier).

pattern DEVICE_ID :: Key Text

A unique identifier representing the device

Notes

The device identifier MUST only be defined using the values outlined below. This value is not an advertising identifier and MUST NOT be used as such. On iOS (Swift or Objective-C), this value MUST be equal to the https://developer.apple.com/documentation/uikit/uidevice/1620059-identifierforvendor. On Android (Java or Kotlin), this value MUST be equal to the Firebase Installation ID or a globally unique UUID which is persisted across sessions in your application. More information can be found https://developer.android.com/training/articles/user-data-ids on best practices and exact implementation details. Caution should be taken when storing personal data or anything which can identify a user. GDPR and data protection laws may apply, ensure you do your own due diligence.

pattern DEVICE_MODEL_IDENTIFIER :: Key Text

The model identifier for the device

Notes

It's recommended this value represents a machine readable version of the model identifier rather than the market or consumer-friendly name of the device.

pattern DEVICE_MODEL_NAME :: Key Text

The marketing name for the device model

Notes

It's recommended this value represents a human readable version of the device model rather than a machine readable alternative.

pattern DEVICE_MANUFACTURER :: Key Text

The name of the device manufacturer

Notes

The Android OS provides this field via https://developer.android.com/reference/android/os/Build#MANUFACTURER. iOS apps SHOULD hardcode the value Apple.

pattern FAAS_NAME :: Key Text

The name of the single function that this runtime instance executes.

Notes

This is the name of the function as configured/deployed on the FaaS platform and is usually different from the name of the callback function (which may be stored in the ../../trace/semantic_conventions/span-general.md#source-code-attributes span attributes).For some cloud providers, the above definition is ambiguous. The following definition of function name MUST be used for this attribute (and consequently the span name) for the listed cloud providersproducts:listrongAzure:<strong> The full name FUNCAPP/FUNC, i.e., function app name followed by a forward slash followed by the function name (this form can also be seen in the resource JSON for the function). This means that a span attribute MUST be used, as an Azure function app can host multiple functions that would usually share a TracerProvider (see also the faas.id attribute)./li

pattern FAAS_ID :: Key Text

The unique ID of the single function that this runtime instance executes.

Notes

On some cloud providers, it may not be possible to determine the full ID at startup, so consider setting faas.id as a span attribute instead.The exact value to use for faas.id depends on the cloud provider:listrongAWS Lambda:/strong The function https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html. Take care not to use the &quot;invoked ARN&quot; directly but replace any https://docs.aws.amazon.com/lambda/latest/dg/configuration-aliases.html with the resolved function version, as the same runtime instance may be invokable with multiple different aliases./li listrongGCP:/strong The https://cloud.google.com/iam/docs/full-resource-names/li listrongAzure:/strong The https://docs.microsoft.com/en-us/rest/api/resources/resources/get-by-id of the invoked function, emnot/em the function app, having the form /subscriptions/SUBSCIPTION_GUID/resourceGroups/RG/providers/Microsoft.Web/sites/FUNCAPP/functions/FUNC. This means that a span attribute MUST be used, as an Azure function app can host multiple functions that would usually share a TracerProvider./li

pattern FAAS_VERSION :: Key Text

The immutable version of the function being executed.

Notes

Depending on the cloud provider and platform, use:listrongAWS Lambda:/strong The https://docs.aws.amazon.com/lambda/latest/dg/configuration-versions.html (an integer represented as a decimal string)./li listrongGoogle Cloud Run:/strong The https://cloud.google.com/run/docs/managing/revisions (i.e., the function name plus the revision suffix)./li listrongGoogle Cloud Functions:/strong The value of the https://cloud.google.com/functions/docs/env-var#runtime_environment_variables_set_automatically./li listrongAzure Functions:/strong Not applicable. Do not set this attribute./li

pattern FAAS_INSTANCE :: Key Text

The execution environment ID as a string, that will be potentially reused for other invocations to the same function/function version.

Notes

listrongAWS Lambda:/strong Use the (full) log stream name./li

pattern FAAS_MAX_MEMORY :: Key Int64

The amount of memory available to the serverless function in MiB.

Notes

It's recommended to set this attribute since e.g. too little memory can easily stop a Java AWS Lambda function from working correctly. On AWS Lambda, the environment variable AWS_LAMBDA_FUNCTION_MEMORY_SIZE provides this information.

pattern HOST_ID :: Key Text

Unique host ID. For Cloud, this must be the instance_id assigned by the cloud provider.

pattern HOST_NAME :: Key Text

Name of the host. On Unix systems, it may contain what the hostname command returns, or the fully qualified hostname, or another name specified by the user.

pattern HOST_TYPE :: Key Text

Type of host. For Cloud, this must be the machine type.

pattern HOST_ARCH :: Key Text

The CPU architecture the host system is running on.

pattern HOST_IMAGE_NAME :: Key Text

Name of the VM image or OS install the host was instantiated from.

pattern HOST_IMAGE_ID :: Key Text

VM image ID. For Cloud, this value is from the provider.

pattern HOST_IMAGE_VERSION :: Key Text

The version string of the VM image as defined in README.md#version-attributes.

pattern K8S_CLUSTER_NAME :: Key Text

The name of the cluster.

pattern K8S_NODE_NAME :: Key Text

The name of the Node.

pattern K8S_NODE_UID :: Key Text

The UID of the Node.

pattern K8S_NAMESPACE_NAME :: Key Text

The name of the namespace that the pod is running in.

pattern K8S_POD_UID :: Key Text

The UID of the Pod.

pattern K8S_POD_NAME :: Key Text

The name of the Pod.

pattern K8S_CONTAINER_NAME :: Key Text

The name of the Container from Pod specification, must be unique within a Pod. Container runtime usually uses different globally unique name (container.name).

pattern K8S_CONTAINER_RESTART_COUNT :: Key Int64

Number of times the container was restarted. This attribute can be used to identify a particular container (running or stopped) within a container spec.

pattern K8S_REPLICASET_UID :: Key Text

The UID of the ReplicaSet.

pattern K8S_REPLICASET_NAME :: Key Text

The name of the ReplicaSet.

pattern K8S_DEPLOYMENT_UID :: Key Text

The UID of the Deployment.

pattern K8S_DEPLOYMENT_NAME :: Key Text

The name of the Deployment.

pattern K8S_STATEFULSET_UID :: Key Text

The UID of the StatefulSet.

pattern K8S_STATEFULSET_NAME :: Key Text

The name of the StatefulSet.

pattern K8S_DAEMONSET_UID :: Key Text

The UID of the DaemonSet.

pattern K8S_DAEMONSET_NAME :: Key Text

The name of the DaemonSet.

pattern K8S_JOB_UID :: Key Text

The UID of the Job.

pattern K8S_JOB_NAME :: Key Text

The name of the Job.

pattern K8S_CRONJOB_UID :: Key Text

The UID of the CronJob.

pattern K8S_CRONJOB_NAME :: Key Text

The name of the CronJob.

pattern OS_TYPE :: Key Text

The operating system type.

pattern OS_DESCRIPTION :: Key Text

Human readable (not intended to be parsed) OS version information, like e.g. reported by ver or lsb_release -a commands.

pattern OS_NAME :: Key Text

Human readable operating system name.

pattern OS_VERSION :: Key Text

The version string of the operating system as defined in ../../resource/semantic_conventions/README.md#version-attributes.

pattern PROCESS_PID :: Key Int64

Process identifier (PID).

pattern PROCESS_EXECUTABLE_NAME :: Key Text

The name of the process executable. On Linux based systems, can be set to the Name in proc/[pid]/status. On Windows, can be set to the base name of GetProcessImageFileNameW.

pattern PROCESS_EXECUTABLE_PATH :: Key Text

The full path to the process executable. On Linux based systems, can be set to the target of proc/[pid]/exe. On Windows, can be set to the result of GetProcessImageFileNameW.

pattern PROCESS_COMMAND :: Key Text

The command used to launch the process (i.e. the command name). On Linux based systems, can be set to the zeroth string in proc/[pid]/cmdline. On Windows, can be set to the first parameter extracted from GetCommandLineW.

pattern PROCESS_COMMAND_LINE :: Key Text

The full command used to launch the process as a single string representing the full command. On Windows, can be set to the result of GetCommandLineW. Do not set this if you have to assemble it just for monitoring; use process.command_args instead.

pattern PROCESS_COMMAND_ARGS :: Key (AttrVals Text)

All the command arguments (including the command/executable itself) as received by the process. On Linux-based systems (and some other Unixoid systems supporting procfs), can be set according to the list of null-delimited strings extracted from proc/[pid]/cmdline. For libc-based executables, this would be the full argv vector passed to main.

pattern PROCESS_OWNER :: Key Text

The username of the user that owns the process.

pattern PROCESS_RUNTIME_NAME :: Key Text

The name of the runtime of this process. For compiled native binaries, this SHOULD be the name of the compiler.

pattern PROCESS_RUNTIME_VERSION :: Key Text

The version of the runtime of this process, as returned by the runtime without modification.

pattern PROCESS_RUNTIME_DESCRIPTION :: Key Text

An additional description about the runtime of the process, for example a specific vendor customization of the runtime environment.

pattern SERVICE_NAME :: Key Text

Logical name of the service.

Notes

MUST be the same for all instances of horizontally scaled services. If the value was not specified, SDKs MUST fallback to unknown_service: concatenated with process.md#process, e.g. unknown_service:bash. If process.executable.name is not available, the value MUST be set to unknown_service.

pattern SERVICE_NAMESPACE :: Key Text

A namespace for service.name.

Notes

A string value having a meaning that helps to distinguish a group of services, for example the team name that owns a group of services. service.name is expected to be unique within the same namespace. If service.namespace is not specified in the Resource then service.name is expected to be unique for all services that have no explicit namespace defined (so the empty/unspecified namespace is simply one more valid namespace). Zero-length namespace string is assumed equal to unspecified namespace.

pattern SERVICE_INSTANCE_ID :: Key Text

The string ID of the service instance.

Notes

MUST be unique for each instance of the same service.namespace,service.name pair (in other words service.namespace,service.name,service.instance.id triplet MUST be globally unique). The ID helps to distinguish instances of the same service that exist at the same time (e.g. instances of a horizontally scaled service). It is preferable for the ID to be persistent and stay the same for the lifetime of the service instance, however it is acceptable that the ID is ephemeral and changes during important lifetime events for the service (e.g. service restarts). If the service has no inherent unique ID that can be used as the value of this attribute it is recommended to generate a random Version 1 or Version 4 RFC 4122 UUID (services aiming for reproducible UUIDs may also use Version 5, see RFC 4122 for more recommendations).

pattern SERVICE_VERSION :: Key Text

The version string of the service API or implementation.

pattern TELEMETRY_SDK_NAME :: Key Text

The name of the telemetry SDK as defined above.

pattern TELEMETRY_SDK_LANGUAGE :: Key Text

The language of the telemetry SDK.

pattern TELEMETRY_SDK_VERSION :: Key Text

The version string of the telemetry SDK.

pattern TELEMETRY_AUTO_VERSION :: Key Text

The version string of the auto instrumentation agent, if used.

pattern WEBENGINE_NAME :: Key Text

The name of the web engine.

pattern WEBENGINE_VERSION :: Key Text

The version of the web engine.

pattern WEBENGINE_DESCRIPTION :: Key Text

Additional description of the web engine (e.g. detailed version and edition information).

data CloudProvider

Constructors

CloudProviderAlibabaCloud
Alibaba Cloud.

CloudProviderAws
Amazon Web Services.

CloudProviderAzure
Microsoft Azure.

CloudProviderGcp
Google Cloud Platform.

CloudProviderTencentCloud
Tencent Cloud.

data CloudPlatform

Constructors

CloudPlatformAlibabaCloudEcs
Alibaba Cloud Elastic Compute Service.

CloudPlatformAlibabaCloudFc
Alibaba Cloud Function Compute.

CloudPlatformAwsEc2
AWS Elastic Compute Cloud.

CloudPlatformAwsEcs
AWS Elastic Container Service.

CloudPlatformAwsEks
AWS Elastic Kubernetes Service.

CloudPlatformAwsLambda
AWS Lambda.

CloudPlatformAwsElasticBeanstalk
AWS Elastic Beanstalk.

CloudPlatformAwsAppRunner
AWS App Runner.

CloudPlatformAzureVm
Azure Virtual Machines.

CloudPlatformAzureContainerInstances
Azure Container Instances.

CloudPlatformAzureAks
Azure Kubernetes Service.

CloudPlatformAzureFunctions
Azure Functions.

CloudPlatformAzureAppService
Azure App Service.

CloudPlatformGcpComputeEngine
Google Cloud Compute Engine (GCE).

CloudPlatformGcpCloudRun
Google Cloud Run.

CloudPlatformGcpKubernetesEngine
Google Cloud Kubernetes Engine (GKE).

CloudPlatformGcpCloudFunctions
Google Cloud Functions (GCF).

CloudPlatformGcpAppEngine
Google Cloud App Engine (GAE).

CloudPlatformTencentCloudCvm
Tencent Cloud Cloud Virtual Machine (CVM).

CloudPlatformTencentCloudEks
Tencent Cloud Elastic Kubernetes Service (EKS).

CloudPlatformTencentCloudScf
Tencent Cloud Serverless Cloud Function (SCF).

data AwsEcsLaunchtype

Constructors

AwsEcsLaunchtypeEc2
ec2.

AwsEcsLaunchtypeFargate
fargate.

data HostArch

Constructors

HostArchAmd64
AMD64.

HostArchArm32
ARM32.

HostArchArm64
ARM64.

HostArchIa64
Itanium.

HostArchPpc32
32-bit PowerPC.

HostArchPpc64
64-bit PowerPC.

HostArchS390x
IBM z/Architecture.

HostArchX86
32-bit x86.

data OsType

Constructors

OsTypeWindows
Microsoft Windows.

OsTypeLinux
Linux.

OsTypeDarwin
Apple Darwin.

OsTypeFreebsd
FreeBSD.

OsTypeNetbsd
NetBSD.

OsTypeOpenbsd
OpenBSD.

OsTypeDragonflybsd
DragonFly BSD.

OsTypeHpux
HP-UX (Hewlett Packard Unix).

OsTypeAix
AIX (Advanced Interactive eXecutive).

OsTypeSolaris
SunOS, Oracle Solaris.

OsTypeZOs
IBM z/OS.

data TelemetrySdkLanguage

Constructors

TelemetrySdkLanguageCpp
cpp.

TelemetrySdkLanguageDotnet
dotnet.

TelemetrySdkLanguageErlang
erlang.

TelemetrySdkLanguageGo
go.

TelemetrySdkLanguageJava
java.

TelemetrySdkLanguageNodejs
nodejs.

TelemetrySdkLanguagePhp
php.

TelemetrySdkLanguagePython
python.

TelemetrySdkLanguageRuby
ruby.

TelemetrySdkLanguageWebjs
webjs.

TelemetrySdkLanguageSwift
swift.