OTel.SDK.Resource.Core.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

data Resource (attrs :: AttrsFor -> Type)

Constructors

Resource

Fields

resourceAttrs :: attrs 'AttrsForResource

resourceSchemaURL :: Maybe SchemaURL

defaultResourceBuilder :: Text -> ResourceBuilder

forSpecificSchema :: SchemaURL -> ResourceBuilder -> ResourceBuilder

resourceBuilderFromAttrs :: AttrsBuilder 'AttrsForResource -> ResourceBuilder

newtype ResourceBuilder

Constructors

ResourceBuilder

Fields

unResourceBuilder :: Either (HashMap SchemaURL (AttrsBuilder 'AttrsForResource)) (Resource AttrsBuilder)

buildResource :: forall m. MonadThrow m => ResourceBuilder -> m (Resource Attrs)

buildResourcePure :: ResourceBuilder -> Either ResourceMergeError (Resource Attrs)

newtype ResourceMergeError

Constructors

ResourceMergeError

Fields

resourceMergeErrorSchemas :: HashMap SchemaURL (Attrs 'AttrsForResource)